requirements.txt
# Enterprise Knowledge Assistant - Dependencies
# Core web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Embeddings and Vector Store
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0

# LLM Integration (Groq - Free tier available)
groq>=0.4.0

# Environment and utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# Optional: OpenAI (if preferred over Groq)
# openai>=1.0.0
